extract step duration calculation to functions package
